Exclude /.vscode directory
This directory is automatically generated by vscode or it's extensions
and are typically of low to no value to a project.
(Especially since extensions usually don't ask for permission)

Review has been skipped due to the size and impact of the change.
PR: https://github.com/jotoho/cpp2-praktikum/pull/15