chore(ci): reduce frequency of thread locking from weekly to monthly
This is a rather small and insignificant repository and you could argue
that having a CI job like lock.yml is nonsense.
It's a resource waste to have it run every week.
After this semester, I may want to disable chronological triggering
completely and only have the job run when manually triggered.